+++ src/funki/app/style.py
@@ -62,3 +62,8 @@ def load_style(root):
     )
 
     return style
+
+styleText = {
+    'background': _colors['lightgray'],
+    'borderwidth': 0,
+}