+++ src/funki/app/utils.py
@@ -305,97 +305,47 @@ class PopUpTable(tk.Toplevel):
                 button_save.configure(state='normal')
 
 
-# TODO: Should probably replace with Text widget, so we can add scrollbar and
-#       also current solution does not wrap text very well
-# Adapted from thegamecracks' gist
-# https://gist.github.com/thegamecracks/5595dad631ec50bdc021f945054e86fb
-class WrapLabel(ttk.Label):
-    '''
-    A label that automatically wraps its text when resized.
+class ScrollText(ttk.Frame):
 
-    When using this label, the geometry manager must be configured
-    to resize the widget in some way, otherwise the width will
-    default to the text's max length.
-    This can mean using ``.pack(fill="both", expand=True)``, or
-    ``.place(relwidth=1, relheight=1)``, or by configuring grid
-    weights appropriately.
+    def __init__(self, parent, text=None, **options):
 
-    ``minwidth=`` can be specified to prevent wrapping under
-    a certain width which can significantly improve performance
-    for longer lines of text.
+        super().__init__(parent)
 
-    As a limitation of the current implementation, this class can only
-    match its bounding box when used with the grid geometry manager.
-    When using other geometry managers like pack or place, WrapLabel will
-    always wrap to the window's entire width, regardless of the actual
-    bounding box allocated for them. As a result, the text may appear clipped.
+        self.columnconfigure(0, weight=1)
+        self.columnconfigure(1, weight=0)
+        self.rowconfigure(0, weight=1)
+        #self.rowconfigure(1, weight=0)
 
-    If you use those geometry managers, it is recommended to grid / pack
-    each WrapLabel inside a frame so it can calculate the wrap length from
-    that frame's width. For example::
+        self.txt = tk.Text(self, **options)
+        self.update(text)
+        self.txt.grid(row=0, column=0, sticky='NSWE')
 
-        # Instead of packing the label directly:
-        label = WrapLabel(root, text="Some long piece of text")
-        label.pack(fill="both", expand=True, padx=200, pady=200)
+        #xscrollbar = ttk.Scrollbar(
+        #    self,
+        #    orient='horizontal',
+        #    command=self.txt.xview,
+        #)
+        #xscrollbar.grid(column=0, row=1, sticky='EW')
+        yscrollbar = ttk.Scrollbar(
+            self,
+            orient='vertical',
+            command=self.txt.yview,
+        )
+        yscrollbar.grid(column=1, row=0, sticky='NS')
 
-        # Pack a container frame and have the label expand inside it:
-        frame = Frame(root)
-        frame.pack(fill="both", expand=True, padx=200, pady=200)
-        label = WrapLabel(frame, text="Some long piece of text")
-        label.pack(fill="both", expand=True)
-
-    In effect, the container frame serves as a way to compute the true
-    bounding box allocated by the geometry manager, helping the label
-    wrap to the correct width.
-    '''
-
-    def __init__(self, *args, minwidth: int = 1, **kwargs):
-
-        super().__init__(*args, **kwargs)
-
-        self.minwidth = minwidth
-        self.bind("<Configure>", self.__on_configure)
+        self.txt.configure(
+            #xscrollcommand=xscrollbar.set,
+            yscrollcommand=yscrollbar.set,
+        )
 
 
-    def __on_configure(self, event: tk.Event):
+    def update(self, text):
 
-        width = max(self.minwidth, self.__get_width())
+        if text:
 
-        if width != 1:  # Prevent wrapping on initial configuration
+            self.txt.configure(state='normal')
 
-            self.configure(wraplength=width)
+            self.txt.delete('0.0', 'end')
+            self.txt.insert('end', text)
 
-
-    def __get_width(self) -> int:
-
-        if self.winfo_manager() == "grid":
-
-            options = self.grid_info()
-            col = int(options.get("column", 0))
-            row = int(options.get("row", 0))
-            colspan = int(options.get("columnspan", 1))
-
-            # Get bbox for the full span (start_col to end_col)
-            bbox = self.master.grid_bbox(col, row, col + colspan, row + 1)
-
-            if bbox is None:
-
-                return 1
-
-            width = bbox[2]
-
-            padx = options.get("padx", 0)
-
-            if isinstance(padx, tuple):
-
-                padx = sum(padx)
-
-            else:
-
-                padx *= 2
-
-            return max(1, width - padx)
-
-        else:
-
-            return self.master.winfo_width()
+            self.txt.configure(state='disabled')