+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # FUNKI changelog
 
 ## 0.2.X
+- Replaced old text wrapping class with scrollable text in a frame
 - Finished implementing storing/loading of all parameters
 - Added change log (this document)